# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/credentials/{credentialId}:
    get:
      operationId: CredentialsController_getCredential
      summary: Retrieve credential
      description: Retrieve a specific Verifiable Credential issued by this wallet
      parameters:
        - name: credentialId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialsDto"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      tags:
        - Credentials
  /api/management/credentials:
    get:
      operationId: CredentialsManagementController_getCredentials
      summary: List credentials
      description: List all credentials, that the current user is allowed to view, in
        this wallet
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CredentialsDto"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      tags: &a7
        - Management Credentials
      security: &a9
        - &a8
          oauth2:
            - wallet_view_all_credentials
            - wallet_view_own_credentials
    post:
      operationId: CredentialsManagementController_addCredential
      summary: Add credential
      description: Issue a new credential within this wallet
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialConfigDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialsDto"
        "400":
          description: Malformed request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "409":
          description: Resource already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      tags: *a7
      security:
        - *a8
        - oauth2:
            - wallet_manage_all_credentials
            - wallet_manage_own_credentials
  /api/management/credentials/config:
    get:
      operationId: CredentialsManagementController_getConfig
      summary: Retrieve credential configuration
      description: Retrieves credential configuration that can be used by this wallet.
        Contains both trust anchors and JSON-LD contexts.
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialConfigDto"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      tags: *a7
      security: *a9
  /api/management/credentials/import:
    post:
      operationId: CredentialsManagementController_importCredential
      summary: Import credential
      description: Import a credential issued by an external credential issuer
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifiableCredentialDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialsDto"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "409":
          description: Resource already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      tags: *a7
      security:
        - *a8
        - oauth2:
            - wallet_manage_all_credentials
            - wallet_manage_own_credentials
  /api/management/credentials/{credentialId}:
    get:
      operationId: CredentialsManagementController_getCredential
      summary: Retrieve credential
      description: Retrieve a specific credential within this wallet
      parameters:
        - name: credentialId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialsDto"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      tags: *a7
      security:
        - *a8
        - oauth2:
            - wallet_manage_all_credentials
            - wallet_manage_own_credentials
    put:
      operationId: CredentialsManagementController_updateCredential
      summary: Update credential
      description: "Update a credential within this wallet. __*Note*__: this will
        either self-issue or import a credential."
      parameters:
        - name: credentialId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialConfigDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialsDto"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      tags: *a7
      security:
        - *a8
        - oauth2:
            - wallet_manage_all_credentials
            - wallet_manage_own_credentials
    delete:
      operationId: CredentialsManagementController_deleteCredential
      summary: Delete credential
      description: Deletes an existing credential within this wallet
      parameters:
        - name: credentialId
          required: true
          in: path
          schema:
            type: string
      responses:
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      tags: *a7
      security:
        - *a8
        - oauth2:
            - wallet_manage_all_credentials
            - wallet_manage_own_credentials
  /api/management/credentials/gaiax/legalRegistrationNumber:
    post:
      operationId: GaiaXManagementController_requestLegalRegistrationNumberCredential
      summary: Issue legal registration number credential
      description: Self-issue a Legal Registration Number credential following the
        Gaia-X Trust Framework
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegalRegistrationNumberRequestDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialsDto"
        "400":
          description: Malformed request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      tags: &a10
        - Management Gaia-X Credentials
      security: &a11
        - oauth2:
            - wallet_manage_all_credentials
            - wallet_manage_own_credentials
  /api/management/credentials/gaiax/compliance:
    post:
      operationId: GaiaXManagementController_requestComplianceCredential
      summary: Request compliance credential
      description: Request a compliance credential from a Gaia-X Digital Clearing
        House based on existing credentials within this wallet
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplianceRequestDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialsDto"
        "400":
          description: Malformed request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      tags: *a10
      security: *a11
...
info:
  title: TSG Wallet
  description: Description
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        password:
          scopes:
            wallet_view_did: wallet_view_did
            wallet_manage_keys: wallet_manage_keys
            wallet_view_own_credentials: wallet_view_own_credentials
            wallet_view_all_credentials: wallet_view_all_credentials
            wallet_manage_own_credentials: wallet_manage_own_credentials
            wallet_manage_all_credentials: wallet_manage_all_credentials
            wallet_issue_credentials: wallet_issue_credentials
            wallet_view_presentations: wallet_view_presentations
  schemas:
...
    ErrorDto:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
        code:
          type: number
        message:
          type: object
        error:
          type: string
      required:
        - name
        - status
        - code
...
    SignatureDto:
      type: object
      properties:
        type:
          type: string
        created:
          type: string
          format: date-time
        proofPurpose:
          type: string
        jws:
          type: string
        verificationMethod:
          type: string
      required:
        - type
        - created
        - proofPurpose
        - jws
        - verificationMethod
    VerifiableCredentialDto:
      type: object
      properties:
        "@context":
          type: array
          items:
            type: string
        type:
          type: array
          items:
            type: string
        id:
          type: string
        credentialSubject:
          oneOf:
            - $ref: "#/components/schemas/DefaultCredentialSubjectDto"
            - type: array
              items:
                $ref: "#/components/schemas/DefaultCredentialSubjectDto"
        issuer:
          type: string
        expirationDate:
          type: string
          format: date-time
        issuanceDate:
          type: string
          format: date-time
        evidence:
          type: object
        proof:
          $ref: "#/components/schemas/SignatureDto"
      required:
        - "@context"
        - type
        - credentialSubject
        - issuer
        - issuanceDate
        - proof
...
    CredentialsDto:
      type: object
      properties:
        id:
          type: string
        targetDid:
          type: string
        credential:
          $ref: "#/components/schemas/VerifiableCredentialDto"
        selfIssued:
          type: boolean
        created:
          format: date-time
          type: string
        modified:
          format: date-time
          type: string
        deleted:
          format: date-time
          type: string
      required:
        - id
        - targetDid
        - credential
        - selfIssued
        - created
        - modified
    DefaultCredentialSubjectDto:
      type: object
      properties:
        id:
          type: string
      required:
        - id
    CredentialConfigDto:
      type: object
      properties:
        context:
          title: JSON-LD Contexts for the credential
          example:
            - https://dataspace.example/context
          type: array
          items:
            type: string
        type:
          title: Credential Type
          example:
            - DataSpaceMembershipCredential
          type: array
          items:
            type: string
        id:
          type: string
          title: Credential ID
        keyId:
          type: string
        credentialSubject:
          $ref: "#/components/schemas/DefaultCredentialSubjectDto"
      required:
        - id
        - credentialSubject
    LegalRegistrationNumberRequestDto:
      type: object
      properties:
        vcId:
          type: string
        clearingHouse:
          type: string
        credentialSubject:
          $ref: "#/components/schemas/DefaultCredentialSubjectDto"
      required:
        - vcId
        - clearingHouse
        - credentialSubject
    ComplianceRequestDto:
      type: object
      properties:
        vcId:
          type: string
        clearingHouse:
          type: string
        credentials:
          type: array
          items:
            $ref: "#/components/schemas/VerifiableCredentialDto"
      required:
        - vcId
        - clearingHouse
        - credentials
...